[Game] {Decorations can be created via console}

+++ centurion-source/assets/shaders/dsprite/vertex.glsl
@@ -10,8 +10,8 @@ uniform mat4 view;
 uniform float x;
 uniform float y;
 uniform float z;
-uniform int w;
-uniform int h;
+uniform float w;
+uniform float h;
 
 void main()
 {
